@@ -32,8 +32,15 @@
   "devDependencies": {
     "@esbuild-plugins/node-globals-polyfill": "^0.1.1",
     "@esbuild-plugins/node-modules-polyfill": "^0.1.4",
-    "@typescript-eslint/eslint-plugin": "^5.52.0",
-    "@typescript-eslint/parser": "^5.52.0",
+    "@eslint/js": "^8.36.0",
+    "@types/connect": "^3.4.35",
+    "@types/morgan": "^1.9.4",
+    "@types/serve-static": "^1.15.1",
+    "@types/sharp": "^0.31.1",
+    "@types/to-ico": "^1.1.1",
+    "@types/ws": "^8.5.4",
+    "@typescript-eslint/eslint-plugin": "^5.55.0",
+    "@typescript-eslint/parser": "^5.55.0",
     "enso-copy-plugin": "^1.0.0",
     "enso-gui-server": "^1.0.0",
     "esbuild": "^0.15.14",
@@ -42,9 +49,10 @@
     "esbuild-plugin-alias": "^0.2.1",
     "esbuild-plugin-time": "^1.0.0",
     "esbuild-plugin-yaml": "^0.0.1",
-    "eslint": "^8.30.0",
-    "eslint-plugin-jsdoc": "^40.0.0",
+    "eslint": "^8.36.0",
+    "eslint-plugin-jsdoc": "^40.0.2",
     "glob": "^8.0.3",
+    "globals": "^13.20.0",
     "source-map-loader": "^4.0.1",
     "ts-loader": "^9.3.0",
     "ts-node": "^10.9.1",
