@@ -3,7 +3,8 @@
  * APIs to the renderer via the contextBridge API. To learn more, visit:
  * https://www.electronjs.org/docs/latest/tutorial/tutorial-preload. */
 
-const { contextBridge, ipcRenderer } = require('electron')
+import * as electron from 'electron'
+
 import * as ipc from 'ipc'
 
 // =================
@@ -22,30 +23,34 @@ const AUTHENTICATION_API_KEY = 'authenticationApi'
 // the primary documentation.
 
 /** Shutdown-related commands and events. */
-contextBridge.exposeInMainWorld('enso_lifecycle', {
+electron.contextBridge.exposeInMainWorld('enso_lifecycle', {
     /** Allows application-exit to be initiated from WASM code.
      * This is used, for example, in a key binding (Ctrl+Alt+Q) that saves a performance profile and
      * exits. */
-    quit: () => ipcRenderer.send(ipc.channel.quit),
+    quit: () => {
+        electron.ipcRenderer.send(ipc.Channel.quit)
+    },
 })
 
 // Save and load profile data.
 let onProfiles: ((profiles: string[]) => void)[] = []
-let profilesLoaded: string[]
-ipcRenderer.on(ipc.channel.profilesLoaded, (event, profiles: string[]) => {
+let profilesLoaded: string[] | undefined
+electron.ipcRenderer.on(ipc.Channel.profilesLoaded, (_event, profiles: string[]) => {
     for (const callback of onProfiles) {
         callback(profiles)
     }
     onProfiles = []
     profilesLoaded = profiles
 })
-contextBridge.exposeInMainWorld('enso_profiling_data', {
+electron.contextBridge.exposeInMainWorld('enso_profiling_data', {
     // Delivers profiling log.
-    saveProfile: (data: any) => ipcRenderer.send(ipc.channel.saveProfile, data),
+    saveProfile: (data: unknown) => {
+        electron.ipcRenderer.send(ipc.Channel.saveProfile, data)
+    },
     // Requests any loaded profiling logs.
     loadProfiles: (callback: (profiles: string[]) => void) => {
         if (profilesLoaded === undefined) {
-            ipcRenderer.send('load-profiles')
+            electron.ipcRenderer.send('load-profiles')
             onProfiles.push(callback)
         } else {
             callback(profilesLoaded)
@@ -54,9 +59,11 @@ contextBridge.exposeInMainWorld('enso_profiling_data', {
 })
 
 // Access to the system console that Electron was run from.
-contextBridge.exposeInMainWorld('enso_console', {
+electron.contextBridge.exposeInMainWorld('enso_console', {
     // Print an error message with `console.error`.
-    error: (data: any) => ipcRenderer.send('error', data),
+    error: (data: unknown) => {
+        electron.ipcRenderer.send('error', data)
+    },
 })
 
 // ==========================
@@ -68,18 +75,19 @@ contextBridge.exposeInMainWorld('enso_console', {
  * - handle deep links from the system browser or email client to the dashboard.
  *
  * Some functions (i.e., the functions to open URLs in the system browser) are not available in
- * sandboxed processes (i.e., the dashboard). So the {@link contextBridge.exposeInMainWorld} API is
+ * sandboxed processes (i.e., the dashboard). So the {@link electron.contextBridge.exposeInMainWorld} API is
  * used to expose these functions. The functions are exposed via this "API object", which is added
  * to the main window.
  *
- * For more details, see: https://www.electronjs.org/docs/latest/api/context-bridge#api-functions */
+ * For more details, see: https://www.electronjs.org/docs/latest/api/context-bridge#api-functions. */
 const AUTHENTICATION_API = {
     /** Opens a URL in the system browser (rather than in the app).
      *
      * OAuth URLs must be opened this way because the dashboard application is sandboxed and thus
      * not privileged to do so unless we explicitly expose this functionality. */
-    openUrlInSystemBrowser: (url: string) =>
-        ipcRenderer.send(ipc.channel.openUrlInSystemBrowser, url),
+    openUrlInSystemBrowser: (url: string) => {
+        electron.ipcRenderer.send(ipc.Channel.openUrlInSystemBrowser, url)
+    },
     /** Set the callback that will be called when a deep link to the application is opened.
      *
      * The callback is intended to handle links like
@@ -87,6 +95,8 @@ const AUTHENTICATION_API = {
      * system browser or email client. Handling the links involves resuming whatever flow was in
      * progress when the link was opened (e.g., an OAuth registration flow). */
     setDeepLinkHandler: (callback: (url: string) => void) =>
-        ipcRenderer.on(ipc.channel.openDeepLink, (_event, url) => callback(url)),
+        electron.ipcRenderer.on(ipc.Channel.openDeepLink, (_event, url: string) => {
+            callback(url)
+        }),
 }
-contextBridge.exposeInMainWorld(AUTHENTICATION_API_KEY, AUTHENTICATION_API)
+electron.contextBridge.exposeInMainWorld(AUTHENTICATION_API_KEY, AUTHENTICATION_API)
