@@ -1,14 +1,21 @@
-/** Script that bundles JS client code. */
+/** @file Script that bundles JS client code. */
 
-import path, { dirname } from 'node:path'
-import esbuild from 'esbuild'
-import { bundlerOptionsFromEnv } from './esbuild-config.js'
-import { fileURLToPath } from 'node:url'
-export const thisPath = path.resolve(dirname(fileURLToPath(import.meta.url)))
+import * as path from 'node:path'
+import * as url from 'node:url'
+
+import * as esbuild from 'esbuild'
+
+import * as esbuildConfig from './esbuild-config.js'
+
+// =================
+// === Constants ===
+// =================
+
+export const THIS_PATH = path.resolve(path.dirname(url.fileURLToPath(import.meta.url)))
 
 // ================
 // === Bundling ===
 // ================
 
-const bundlerOptions: esbuild.BuildOptions = bundlerOptionsFromEnv()
-await esbuild.build(bundlerOptions)
+const BUNDLER_OPTIONS: esbuild.BuildOptions = esbuildConfig.bundlerOptionsFromEnv()
+await esbuild.build(BUNDLER_OPTIONS)
