@@ -6,31 +6,32 @@
  * - The default configuration of the build process. */
 /** @module */
 
-import path from 'node:path'
-import child_process from 'node:child_process'
-import fs from 'node:fs/promises'
-import { CliOptions, Configuration, LinuxTargetSpecificOptions, Platform } from 'electron-builder'
-import builder from 'electron-builder'
-import { notarize } from 'electron-notarize'
+import * as childProcess from 'node:child_process'
+import * as fs from 'node:fs/promises'
+import * as path from 'node:path'
+
+import * as electronBuilder from 'electron-builder'
+import * as electronNotarize from 'electron-notarize'
+import * as macOptions from 'app-builder-lib/out/options/macOptions'
+import yargs from 'yargs'
+
+import * as paths from './paths.js'
+import * as shared from './shared.js'
 import signArchivesMacOs from './tasks/signArchivesMacOs.js'
 
-import { project_manager_bundle } from './paths.js'
-import * as shared from './shared.js'
-import build from '../../build.json' assert { type: 'json' }
-import yargs from 'yargs'
-import { MacOsTargetName } from 'app-builder-lib/out/options/macOptions'
+import BUILD_INFO from '../../build.json' assert { type: 'json' }
 
 /** The parts of the electron-builder configuration that we want to keep configurable.
  *
  * @see `args` definition below for fields description.
- * */
+ */
 export interface Arguments {
     target?: string
     iconsDist: string
     guiDist: string
     ideDist: string
     projectManagerDist: string
-    platform: Platform
+    platform: electronBuilder.Platform
 }
 
 /** CLI argument parser (with support for environment variables) that provides the necessary options. */
@@ -65,8 +66,8 @@ export const args: Arguments = await yargs(process.argv.slice(2))
         platform: {
             type: 'string',
             description: 'Platform that Electron Builder should target',
-            default: Platform.current().toString(),
-            coerce: (p: string) => Platform.fromString(p),
+            default: electronBuilder.Platform.current().toString(),
+            coerce: (p: string) => electronBuilder.Platform.fromString(p),
         },
         target: {
             type: 'string',
@@ -75,16 +76,16 @@ export const args: Arguments = await yargs(process.argv.slice(2))
     }).argv
 
 /** Based on the given arguments, creates a configuration for the Electron Builder. */
-export function createElectronBuilderConfig(args: Arguments): Configuration {
+export function createElectronBuilderConfig(passedArgs: Arguments): electronBuilder.Configuration {
     return {
         appId: 'org.enso',
         productName: shared.PRODUCT_NAME,
         extraMetadata: {
-            version: build.version,
+            version: BUILD_INFO.version,
         },
         copyright: 'Copyright © 2022 ${author}.',
         artifactName: 'enso-${os}-${version}.${ext}',
-        /** Definitions of URL {@link builder.Protocol} schemes used by the IDE.
+        /** Definitions of URL {@link electronBuilder.Protocol} schemes used by the IDE.
          *
          * Electron will register all URL protocol schemes defined here with the OS. Once a URL protocol
          * scheme is registered with the OS, any links using that scheme will function as "deep links".
@@ -97,7 +98,7 @@ export function createElectronBuilderConfig(args: Arguments): Configuration {
          * - navigate to the location specified by the URL of the deep link.
          *
          * For details on how this works, see:
-         * https://www.electronjs.org/docs/latest/tutorial/launch-app-from-url-in-another-app */
+         * https://www.electronjs.org/docs/latest/tutorial/launch-app-from-url-in-another-app. */
         protocols: [
             /** Electron URL protocol scheme definition for deep links to authentication flow pages. */
             {
@@ -108,8 +109,8 @@ export function createElectronBuilderConfig(args: Arguments): Configuration {
         ],
         mac: {
             // We do not use compression as the build time is huge and file size saving is almost zero.
-            target: (args.target as MacOsTargetName) ?? 'dmg',
-            icon: `${args.iconsDist}/icon.icns`,
+            target: (passedArgs.target ?? 'dmg') as macOptions.MacOsTargetName,
+            icon: `${passedArgs.iconsDist}/icon.icns`,
             category: 'public.app-category.developer-tools',
             darkModeSupport: true,
             type: 'distribution',
@@ -126,24 +127,24 @@ export function createElectronBuilderConfig(args: Arguments): Configuration {
         },
         win: {
             // We do not use compression as the build time is huge and file size saving is almost zero.
-            target: args.target ?? 'nsis',
-            icon: `${args.iconsDist}/icon.ico`,
+            target: passedArgs.target ?? 'nsis',
+            icon: `${passedArgs.iconsDist}/icon.ico`,
         },
         linux: {
             // We do not use compression as the build time is huge and file size saving is almost zero.
-            target: args.target ?? 'AppImage',
-            icon: `${args.iconsDist}/png`,
+            target: passedArgs.target ?? 'AppImage',
+            icon: `${passedArgs.iconsDist}/png`,
             category: 'Development',
         },
         files: [
             '!**/node_modules/**/*',
-            { from: `${args.guiDist}/`, to: '.' },
-            { from: `${args.ideDist}/client`, to: '.' },
+            { from: `${passedArgs.guiDist}/`, to: '.' },
+            { from: `${passedArgs.ideDist}/client`, to: '.' },
         ],
         extraResources: [
             {
-                from: `${args.projectManagerDist}/`,
-                to: project_manager_bundle,
+                from: `${passedArgs.projectManagerDist}/`,
+                to: paths.PROJECT_MANAGER_BUNDLE,
                 filter: ['!**.tar.gz', '!**.zip'],
             },
         ],
@@ -155,7 +156,7 @@ export function createElectronBuilderConfig(args: Arguments): Configuration {
             },
         ],
         directories: {
-            output: `${args.ideDist}`,
+            output: `${passedArgs.ideDist}`,
         },
         nsis: {
             // Disables "block map" generation during electron building. Block maps
@@ -186,47 +187,57 @@ export function createElectronBuilderConfig(args: Arguments): Configuration {
         afterAllArtifactBuild: path.join('tasks', 'computeHashes.cjs'),
 
         afterPack: ctx => {
-            if (args.platform === Platform.MAC) {
+            if (passedArgs.platform === electronBuilder.Platform.MAC) {
                 // Make the subtree writable, so we can sign the binaries.
                 // This is needed because GraalVM distribution comes with read-only binaries.
-                child_process.execFileSync('chmod', ['-R', 'u+w', ctx.appOutDir])
+                childProcess.execFileSync('chmod', ['-R', 'u+w', ctx.appOutDir])
             }
         },
 
         afterSign: async context => {
             // Notarization for macOS.
-            if (args.platform === Platform.MAC && process.env['CSC_LINK']) {
-                const { packager, appOutDir } = context
-                const appName = packager.appInfo.productFilename
+            if (passedArgs.platform === electronBuilder.Platform.MAC && process.env.CSC_LINK) {
+                const {
+                    packager: {
+                        // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+                        platformSpecificBuildOptions: buildOptions,
+                        appInfo: { productFilename: appName },
+                        config: { mac: macConfig },
+                    },
+                    appOutDir,
+                } = context
 
                 // We need to manually re-sign our build artifacts before notarization.
                 console.log('  • Performing additional signing of dependencies.')
                 await signArchivesMacOs({
                     appOutDir: appOutDir,
                     productFilename: appName,
-                    // @ts-ignore
-                    entitlements: context.packager.config.mac.entitlements,
+                    // This will always be defined since we have an `entitlements.mac.plist`.
+                    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+                    entitlements: macConfig!.entitlements!,
                     identity: 'Developer ID Application: New Byte Order Sp. z o. o. (NM77WTZJFQ)',
                 })
 
                 console.log('  • Notarizing.')
-                await notarize({
-                    appBundleId: packager.platformSpecificBuildOptions.appId,
+                await electronNotarize.notarize({
+                    // This will always be defined since we set it at the top of this object.
+                    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+                    appBundleId: (buildOptions as macOptions.MacConfiguration).appId!,
                     appPath: `${appOutDir}/${appName}.app`,
-                    // @ts-ignore
                     appleId: process.env.APPLEID,
-                    // @ts-ignore
                     appleIdPassword: process.env.APPLEIDPASS,
                 })
             }
         },
 
+        // Third-party API specifies `null`, not `undefined`.
+        // eslint-disable-next-line no-restricted-syntax
         publish: null,
     }
 }
 
 /** Build the IDE package with Electron Builder. */
-export async function buildPackage(args: Arguments) {
+export async function buildPackage(passedArgs: Arguments) {
     // `electron-builder` checks for presence of `node_modules` directory. If it is not present, it will
     // install dependencies with `--production` flag (erasing all dev-only dependencies). This does not
     // work sensibly with NPM workspaces. We have our `node_modules` in the root directory, not here.
@@ -235,11 +246,11 @@ export async function buildPackage(args: Arguments) {
     // because of that.
     await fs.mkdir('node_modules', { recursive: true })
 
-    const cli_opts: CliOptions = {
-        config: createElectronBuilderConfig(args),
-        targets: args.platform.createTarget(),
+    const cliOpts: electronBuilder.CliOptions = {
+        config: createElectronBuilderConfig(passedArgs),
+        targets: passedArgs.platform.createTarget(),
     }
-    console.log('Building with configuration:', cli_opts)
-    const result = await builder.build(cli_opts)
+    console.log('Building with configuration:', cliOpts)
+    const result = await electronBuilder.build(cliOpts)
     console.log('Electron Builder is done. Result:', result)
 }
