@@ -1,18 +1,27 @@
-import child_process, { SpawnOptions } from 'child_process'
+/** @file Project Manager bindings. */
+
+import * as childProcess from 'node:child_process'
+import * as fsSync from 'node:fs'
+import * as util from 'node:util'
+
+import * as contentConfig from 'enso-content-config'
+
 import * as config from 'config'
-import fss from 'node:fs'
-import { logger } from 'enso-content-config'
-import util from 'node:util'
-const execFile = util.promisify(child_process.execFile)
+
+const logger = contentConfig.logger
+// This is a wrapped function, so it should be `camelCase`.
+// eslint-disable-next-line no-restricted-syntax
+const execFile = util.promisify(childProcess.execFile)
 
 // =======================
 // === Project Manager ===
 // =======================
 
-/** Return the Project Manager path if it is valid. Otherwise, throw an error. */
+/** Return the Project Manager path.
+ * @throws If the Project Manager path is invalid. */
 export function pathOrPanic(args: config.Args): string {
-    let binPath = args.groups.engine.options.projectManagerPath.value
-    let binExists = fss.existsSync(binPath)
+    const binPath = args.groups.engine.options.projectManagerPath.value
+    const binExists = fsSync.existsSync(binPath)
     if (!binExists) {
         throw new Error(`Could not find the project manager binary at ${binPath}.`)
     }
@@ -21,7 +30,7 @@ export function pathOrPanic(args: config.Args): string {
 
 /** Executes the Project Manager with given arguments. */
 async function exec(args: config.Args, processArgs: string[]) {
-    let binPath = pathOrPanic(args)
+    const binPath = pathOrPanic(args)
     return await execFile(binPath, processArgs).catch(function (err) {
         throw err
     })
@@ -32,23 +41,24 @@ async function exec(args: config.Args, processArgs: string[]) {
  * The standard output and error handles will be redirected to the electron's app output and error
  * handles. Input is piped to this process, so it will not be closed, until this process
  * finished. */
-export function spawn(args: config.Args, processArgs: string[]): child_process.ChildProcess {
+export function spawn(args: config.Args, processArgs: string[]): childProcess.ChildProcess {
     return logger.groupMeasured(
-        `Starting the backend process with the following options: ${processArgs}`,
+        `Starting the backend process with the following options: ${processArgs.join(', ')}.`,
         () => {
             const binPath = pathOrPanic(args)
-            const stdin = 'pipe' as const
-            const stdout = 'inherit' as const
-            const stderr = 'inherit' as const
-            const stdio = [stdin, stdout, stderr]
-            const process = child_process.spawn(binPath, processArgs, { stdio })
-            logger.log(`Backend has been spawned (pid = ${process.pid}).`)
-            process.on('exit', code => logger.log(`Backend exited with code ${code}.`))
+            const process = childProcess.spawn(binPath, processArgs, {
+                stdio: [/* stdin */ 'pipe', /* stdout */ 'inherit', /* stderr */ 'inherit'],
+            })
+            logger.log(`Backend has been spawned (pid = ${String(process.pid)}).`)
+            process.on('exit', code => {
+                logger.log(`Backend exited with code ${String(code)}.`)
+            })
             return process
         }
     )
 }
 
+/** Get the Project Manager version. */
 export async function version(args: config.Args) {
     if (args.options.engine.value) {
         return await exec(args, ['--version']).then(t => t.stdout)
