@@ -1,20 +1,22 @@
 /** @file A simple HTTP server which serves application data to the Electron web-view. */
 
-// @ts-expect-error
-import createServer from 'create-servers'
-import * as fs from 'fs'
+import * as fs from 'node:fs'
+import * as http from 'node:http'
+import * as path from 'node:path'
+
 import * as mime from 'mime-types'
-import * as path from 'path'
 import * as portfinder from 'portfinder'
-import { logger } from 'enso-content-config'
+import createServer from 'create-servers'
+
+import * as contentConfig from 'enso-content-config'
+
+const logger = contentConfig.logger
 
 // =================
 // === Constants ===
 // =================
 
-const responses = {
-    ok: 200,
-}
+const HTTP_STATUS_OK = 200
 
 // ======================
 // === URL Parameters ===
@@ -22,23 +24,29 @@ const responses = {
 
 /** Construct URL query with the given parameters. For each `key` - `value` pair,
  * `key=value` will be added to the query. */
-export function urlParamsFromObject(obj: { [key: string]: string }) {
-    let params = []
+export function urlParamsFromObject(obj: Record<string, string>) {
+    const params = []
     for (const [key, value] of Object.entries(obj)) {
         params.push(`${key}=${encodeURIComponent(value)}`)
     }
-    return params.length == 0 ? '' : '?' + params.join('&')
+    return params.length === 0 ? '' : '?' + params.join('&')
 }
 
 // ==============
 // === Config ===
 // ==============
 
+/** Constructor parameter for the server configuration. */
+interface ConfigConfig {
+    dir: string
+    port: number
+}
+
 /** Server configuration. */
 export class Config {
     dir: string
     port: number
-    constructor(cfg: { dir: string; port: number }) {
+    constructor(cfg: ConfigConfig) {
         this.dir = path.resolve(cfg.dir)
         this.port = cfg.port
     }
@@ -63,17 +71,14 @@ async function findPort(port: number): Promise<number> {
 /// Initially it was based on `union`, but later we migrated to `create-servers`. Read this topic to
 /// learn why: https://github.com/http-party/http-server/issues/483
 export class Server {
-    config: Config
-    server: any
-    constructor(config: Config) {
-        this.config = config
-    }
+    server: unknown
+    constructor(public config: Config) {}
 
     /** Server constructor. */
     static async create(config: Config): Promise<Server> {
-        let local_config = Object.assign({}, config)
-        local_config.port = await findPort(local_config.port)
-        const server = new Server(local_config)
+        const localConfig = Object.assign({}, config)
+        localConfig.port = await findPort(localConfig.port)
+        const server = new Server(localConfig)
         await server.run()
         return server
     }
@@ -85,32 +90,39 @@ export class Server {
                     http: this.config.port,
                     handler: this.process.bind(this),
                 },
-                (err: any) => {
+                err => {
                     if (err) {
                         logger.error(`Error creating server:`, err.http)
                         reject(err)
                     }
                     logger.log(`Server started on port ${this.config.port}.`)
-                    logger.log(`Serving files from '${process.cwd()}/${this.config.dir}'.`)
+                    logger.log(`Serving files from '${path.join(process.cwd(), this.config.dir)}'.`)
                     resolve()
                 }
             )
         })
     }
 
-    process(request: { url: string }, response: any) {
-        const url = request.url.split('?')[0]
-        const resource = url == '/' ? '/index.html' : request.url
-        let resource_file = `${this.config.dir}${resource}`
-        fs.readFile(resource_file, (err: any, data: any) => {
+    process(request: http.IncomingMessage, response: http.ServerResponse) {
+        const requestUrl = request.url
+        if (requestUrl === undefined) {
+            logger.error('Request URL is null.')
+            return
+        }
+        const url = requestUrl.split('?')[0]
+        const resource = url === '/' ? '/index.html' : requestUrl
+        const resourceFile = `${this.config.dir}${resource}`
+        fs.readFile(resourceFile, (err, data) => {
             if (err) {
                 logger.error(`Resource '${resource}' not found.`)
             } else {
-                let contentType = mime.contentType(path.extname(resource_file))
-                let contentLength = data.length
-                response.setHeader('Content-Type', contentType)
+                const contentType = mime.contentType(path.extname(resourceFile))
+                const contentLength = data.length
+                if (contentType !== false) {
+                    response.setHeader('Content-Type', contentType)
+                }
                 response.setHeader('Content-Length', contentLength)
-                response.writeHead(responses.ok)
+                response.writeHead(HTTP_STATUS_OK)
                 response.end(data)
             }
         })
