@@ -1,23 +1,30 @@
 /** @file Application debug information. */
 
-import buildCfg from '../../../build.json'
+import BUILD_INFO from '../../../build.json' assert { type: 'json' }
+
+// =================
+// === Constants ===
+// =================
+
+const INDENT_SIZE = 4
 
 // ==================
 // === Debug Info ===
 // ==================
 
 /** Information about versions of different application components. */
-export const versionInfo = {
-    version: buildCfg.version,
-    build: buildCfg.commit,
+export const VERSION_INFO = {
+    version: BUILD_INFO.version,
+    build: BUILD_INFO.commit,
     electron: process.versions.electron,
     chrome: process.versions.chrome,
 }
 
+/** Get the current system information, useful for debugging. */
 async function getInfo() {
-    let procMemInfo = await process.getProcessMemoryInfo()
+    const procMemInfo = await process.getProcessMemoryInfo()
     return {
-        version: versionInfo,
+        version: VERSION_INFO,
         time: {
             current: Date.now(),
             creation: process.getCreationTime(),
@@ -40,7 +47,8 @@ async function getInfo() {
     }
 }
 
+/** Print the current system information. */
 export async function printInfo() {
-    let info = await getInfo()
-    console.log(JSON.stringify(info, undefined, 4))
+    const info = await getInfo()
+    console.log(JSON.stringify(info, undefined, INDENT_SIZE))
 }
