@@ -6,16 +6,16 @@
 // ===============
 
 /** Channel names used by the IPC protocol. */
-export const channel = {
-    error: 'error',
-    loadProfiles: 'load-profiles',
-    profilesLoaded: 'profiles-loaded',
-    saveProfile: 'save-profile',
-    quit: 'quit-ide',
+export enum Channel {
+    error = 'error',
+    loadProfiles = 'load-profiles',
+    profilesLoaded = 'profiles-loaded',
+    saveProfile = 'save-profile',
+    quit = 'quit-ide',
     /** Channel for requesting that a URL be opened by the system browser. */
-    openUrlInSystemBrowser: 'open-url-in-system-browser',
+    openUrlInSystemBrowser = 'open-url-in-system-browser',
     /** Channel for setting a callback that handles deep links to this application. */
-    setDeepLinkHandler: 'set-deep-link-handler',
+    setDeepLinkHandler = 'set-deep-link-handler',
     /** Channel for signaling that a deep link to this application was opened. */
-    openDeepLink: 'open-deep-link',
+    openDeepLink = 'open-deep-link',
 }
