@@ -1,17 +1,17 @@
 /** @file Security configuration of Electron. Most of the options are based on the official security
- * guide: https://www.electronjs.org/docs/latest/tutorial/security */
+ * guide: https://www.electronjs.org/docs/latest/tutorial/security. */
 
-import Electron from 'electron'
+import * as electron from 'electron'
 
 // =================
 // === Constants ===
 // =================
 
 /** The list of hosts that the app can access. They are required for user authentication to work. */
-const trustedHosts = ['accounts.google.com', 'accounts.youtube.com', 'github.com']
+const TRUSTED_HOSTS = ['accounts.google.com', 'accounts.youtube.com', 'github.com']
 
 /** The list of URLs a new WebView can be pointed to. */
-const webViewUrlWhitelist: string[] = []
+const WEBVIEW_URL_WHITELIST: string[] = []
 
 // =============
 // === Utils ===
@@ -19,7 +19,7 @@ const webViewUrlWhitelist: string[] = []
 
 /** Secure the web preferences of a new window. It deletes potentially unsecure options, making them
  * revert to secure defaults. */
-export function secureWebPreferences(webPreferences: Electron.WebPreferences) {
+export function secureWebPreferences(webPreferences: electron.WebPreferences) {
     delete webPreferences.preload
     delete webPreferences.nodeIntegration
     delete webPreferences.nodeIntegrationInWorker
@@ -36,36 +36,34 @@ export function secureWebPreferences(webPreferences: Electron.WebPreferences) {
 // ================
 
 /** Enabling sandbox globally. Follow the link to learn more:
- * https://www.electronjs.org/docs/latest/tutorial/sandbox */
+ * https://www.electronjs.org/docs/latest/tutorial/sandbox. */
 function enableGlobalSandbox() {
-    Electron.app.enableSandbox()
+    electron.app.enableSandbox()
 }
 
 /** By default, Electron will automatically approve all permission requests unless the developer has
- *  manually configured a custom handler. While a solid default, security-conscious developers might
- *  want to assume the very opposite. Follow the link to learn more:
- *  https://www.electronjs.org/docs/latest/tutorial/security#5-handle-session-permission-requests-from-remote-content */
+ * manually configured a custom handler. While a solid default, security-conscious developers might
+ * want to assume the very opposite. Follow the link to learn more:
+ * https://www.electronjs.org/docs/latest/tutorial/security#5-handle-session-permission-requests-from-remote-content. */
 function rejectPermissionRequests() {
-    Electron.app.whenReady().then(() => {
-        Electron.session.defaultSession.setPermissionRequestHandler(
-            (webContents, permission, callback) => {
-                console.error(`Unhandled permission request '${permission}'.`)
-            }
-        )
+    void electron.app.whenReady().then(() => {
+        electron.session.defaultSession.setPermissionRequestHandler((_webContents, permission) => {
+            console.error(`Unhandled permission request '${permission}'.`)
+        })
     })
 }
 
 /** A WebView created in a renderer process that does not have Node.js integration enabled will not
- *  be able to enable integration itself. However, a WebView will always create an independent
- *  renderer process with its own webPreferences. It is a good idea to control the creation of new
- *  <webview> tags from the main process and to verify that their webPreferences do not disable
- *  security features. Follow the link to learn more:
- *  https://www.electronjs.org/docs/tutorial/security#11-verify-webview-options-before-creation */
+ * be able to enable integration itself. However, a WebView will always create an independent
+ * renderer process with its own webPreferences. It is a good idea to control the creation of new
+ * <webview> tags from the main process and to verify that their webPreferences do not disable
+ * security features. Follow the link to learn more:
+ * https://www.electronjs.org/docs/tutorial/security#11-verify-webview-options-before-creation. */
 function limitWebViewCreation() {
-    Electron.app.on('web-contents-created', (event, contents) => {
+    electron.app.on('web-contents-created', (_event, contents) => {
         contents.on('will-attach-webview', (event, webPreferences, params) => {
             secureWebPreferences(webPreferences)
-            if (params.src && !webViewUrlWhitelist.includes(params.src)) {
+            if (params.src && !WEBVIEW_URL_WHITELIST.includes(params.src)) {
                 console.error(`Blocked the creation of WebView pointing to '${params.src}'`)
                 event.preventDefault()
             }
@@ -76,12 +74,12 @@ function limitWebViewCreation() {
 /** Navigation is a common attack vector. If an attacker can convince your app to navigate away from
  * its current page, they can possibly force your app to open web sites on the Internet. Follow the
  * link to learn more:
- * https://www.electronjs.org/docs/tutorial/security#12-disable-or-limit-navigation */
+ * https://www.electronjs.org/docs/tutorial/security#12-disable-or-limit-navigation. */
 function preventNavigation() {
-    Electron.app.on('web-contents-created', (event, contents) => {
+    electron.app.on('web-contents-created', (_event, contents) => {
         contents.on('will-navigate', (event, navigationUrl) => {
             const parsedUrl = new URL(navigationUrl)
-            if (parsedUrl.origin !== origin && !trustedHosts.includes(parsedUrl.host)) {
+            if (parsedUrl.origin !== origin && !TRUSTED_HOSTS.includes(parsedUrl.host)) {
                 event.preventDefault()
                 console.error(`Prevented navigation to '${navigationUrl}'.`)
             }
@@ -90,12 +88,12 @@ function preventNavigation() {
 }
 
 /** Much like navigation, the creation of new webContents is a common attack vector. Attackers
- *  attempt to convince your app to create new windows, frames, or other renderer processes with
- *  more privileges than they had before or with pages opened that they couldn't open before.
- *  Follow the link to learn more:
- *  https://www.electronjs.org/docs/tutorial/security#13-disable-or-limit-creation-of-new-windows */
+ * attempt to convince your app to create new windows, frames, or other renderer processes with
+ * more privileges than they had before or with pages opened that they couldn't open before.
+ * Follow the link to learn more:
+ * https://www.electronjs.org/docs/tutorial/security#13-disable-or-limit-creation-of-new-windows. */
 function disableNewWindowsCreation() {
-    Electron.app.on('web-contents-created', (event, contents) => {
+    electron.app.on('web-contents-created', (_event, contents) => {
         contents.setWindowOpenHandler(({ url }) => {
             console.error(`Blocking new window creation request to '${url}'.`)
             return { action: 'deny' }
