@@ -1,23 +1,20 @@
-/**
- This script signs the content of all archives that we have for macOS. For this to work this needs
- to run on macOS with `codesign`, and a JDK installed. `codesign` is needed to sign the files,
- while the JDK is needed for correct packing and unpacking of java archives.
+/** @file This script signs the content of all archives that we have for macOS. For this to work this needs to run on
+ * macOS with `codesign`, and a JDK installed. `codesign` is needed to sign the files, while the JDK is needed for
+ * correct packing and unpacking of java archives.
+ *
+ * We require this extra step as our dependencies contain files that require us to re-sign jar contents that cannot be
+ * opened as pure zip archives, but require a java toolchain to extract and re-assemble to preserve manifest
+ * information. This functionality is not provided by `electron-osx-sign` out of the box.
+ *
+ * This code is based on https://github.com/electron/electron-osx-sign/pull/231 but our use-case is unlikely to be
+ * supported by electron-osx-sign as it adds a java toolchain as additional dependency.
+ * This script should be removed once the engine is signed. */
 
- We require this extra step as our dependencies contain files that require us to re-sign jar
- contents that cannot be opened as pure zip archives, but require a java toolchain to extract
- and re-assemble to preserve manifest information. This functionality is not provided by
- `electron-osx-sign` out of the box.
+import * as childProcess from 'node:child_process'
+import * as fs from 'node:fs/promises'
+import * as os from 'node:os'
+import * as pathModule from 'node:path'
 
- This code is based on https://github.com/electron/electron-osx-sign/pull/231 but our use-case
- is unlikely to be supported by electron-osx-sign as it adds a java toolchain as additional
- dependency.
- This script should be removed once the engine is signed.
-**/
-
-import fs from 'node:fs/promises'
-import os from 'node:os'
-import path from 'node:path'
-import child_process from 'node:child_process'
 import glob from 'fast-glob'
 
 // ===============================================
@@ -71,7 +68,7 @@ async function graalSignables(resourcesDir: string): Promise<Signable[]> {
 
     // We use `*` for Graal versioned directory to not have to update this script on every GraalVM update.
     // Updates might still be needed when the list of binaries to sign changes.
-    const graalDir = path.join(resourcesDir, 'enso', 'runtime', '*')
+    const graalDir = pathModule.join(resourcesDir, 'enso', 'runtime', '*')
     const archives = await ArchiveToSign.lookupMany(graalDir, archivePatterns)
     const binaries = await BinaryToSign.lookupMany(graalDir, binariesPatterns)
     return [...archives, ...binaries]
@@ -121,16 +118,16 @@ interface SigningContext {
 /** An entity that we want to sign. */
 interface Signable {
     /** Sign this entity. */
-    sign(context: SigningContext): Promise<void>
+    sign: (context: SigningContext) => Promise<void>
 }
 
 /** Placeholder name for temporary archives. */
-const tmpArchive = 'temporary_archive.zip'
+const TEMPORARY_ARCHIVE_PATH = 'temporary_archive.zip'
 
 /** Helper to execute a program in a given directory and return the output. */
-const run = (cmd: string, args: string[], cwd?: string) => {
+function run(cmd: string, args: string[], cwd?: string) {
     console.log('Running', cmd, args, cwd)
-    return child_process.execFileSync(cmd, args, { cwd }).toString()
+    return childProcess.execFileSync(cmd, args, { cwd }).toString()
 }
 
 /** Archive with some binaries that we want to sign.
@@ -138,28 +135,28 @@ const run = (cmd: string, args: string[], cwd?: string) => {
  * Can be either a zip or a jar file.
  */
 class ArchiveToSign implements Signable {
-    /** An absolute path to the archive. */
-    path: string
-
-    /** A list of patterns for files to sign inside the archive.
-     *
-     *  Relative to the root of the archive.
-     */
-    binaries: glob.Pattern[]
+    /** Looks up for archives to sign using the given path patterns. */
+    static lookupMany = lookupManyHelper(ArchiveToSign.lookup.bind(this))
 
     /** Create a new instance. */
-    constructor(path: string, binaries: glob.Pattern[]) {
-        this.path = path
-        this.binaries = binaries
+    constructor(
+        /** An absolute path to the archive. */
+        public path: string,
+        /** A list of patterns for files to sign inside the archive.
+         * Relative to the root of the archive. */
+        public binaries: glob.Pattern[]
+    ) {}
+
+    /** Looks up for archives to sign using the given path pattern. */
+    static async lookup(base: string, [pattern, binaries]: ArchivePattern) {
+        return lookupHelper(path => new ArchiveToSign(path, binaries))(base, pattern)
     }
 
-    /**
-     * Sign content of an archive. This function extracts the archive, signs the required files,
-     * re-packages the archive and replaces the original.
-     */
+    /** Sign content of an archive. This function extracts the archive, signs the required files,
+     * re-packages the archive and replaces the original. */
     async sign(context: SigningContext) {
         console.log(`Signing archive ${this.path}`)
-        const archiveName = path.basename(this.path)
+        const archiveName = pathModule.basename(this.path)
         const workingDir = await getTmpDir()
         try {
             const isJar = archiveName.endsWith(`jar`)
@@ -175,21 +172,27 @@ class ArchiveToSign implements Signable {
 
             const binariesToSign = await BinaryToSign.lookupMany(workingDir, this.binaries)
             for (const binaryToSign of binariesToSign) {
-                binaryToSign.sign(context)
+                void binaryToSign.sign(context)
             }
 
             if (isJar) {
                 if (archiveName.includes(`runner`)) {
-                    run(`jar`, ['-cfm', tmpArchive, 'META-INF/MANIFEST.MF', '.'], workingDir)
+                    run(
+                        `jar`,
+                        ['-cfm', TEMPORARY_ARCHIVE_PATH, 'META-INF/MANIFEST.MF', '.'],
+                        workingDir
+                    )
                 } else {
-                    run(`jar`, ['-cf', tmpArchive, '.'], workingDir)
+                    run(`jar`, ['-cf', TEMPORARY_ARCHIVE_PATH, '.'], workingDir)
                 }
             } else {
-                run(`zip`, ['-rm', tmpArchive, '.'], workingDir)
+                run(`zip`, ['-rm', TEMPORARY_ARCHIVE_PATH, '.'], workingDir)
             }
 
             // We cannot use fs.rename because temp and target might be on different volumes.
-            console.log(run(`/bin/mv`, [path.join(workingDir, tmpArchive), this.path]))
+            console.log(
+                run(`/bin/mv`, [pathModule.join(workingDir, TEMPORARY_ARCHIVE_PATH), this.path])
+            )
             console.log(
                 `Successfully repacked ${this.path} to handle signing inner native dependency.`
             )
@@ -205,25 +208,21 @@ class ArchiveToSign implements Signable {
             await rmRf(workingDir)
         }
     }
-
-    /** Looks up for archives to sign using the given path pattern. */
-    static async lookup(base: string, [pattern, binaries]: ArchivePattern) {
-        return lookupHelper(path => new ArchiveToSign(path, binaries))(base, pattern)
-    }
-
-    /** Looks up for archives to sign using the given path patterns. */
-    static lookupMany = lookupManyHelper(ArchiveToSign.lookup)
 }
 
 /** A single code binary file to be signed. */
 class BinaryToSign implements Signable {
-    /** An absolute path to the binary. */
-    path: string
+    /** Looks up for binaries to sign using the given path pattern. */
+    static lookup = lookupHelper(path => new BinaryToSign(path))
+
+    /** Looks up for binaries to sign using the given path patterns. */
+    static lookupMany = lookupManyHelper(BinaryToSign.lookup)
 
     /** Create a new instance. */
-    constructor(path: string) {
-        this.path = path
-    }
+    constructor(
+        /** An absolute path to the binary. */
+        public path: string
+    ) {}
 
     /** Sign this binary. */
     async sign({ entitlements, identity }: SigningContext) {
@@ -238,13 +237,9 @@ class BinaryToSign implements Signable {
             identity,
             this.path,
         ])
+        // Async functions should contain await.
+        await Promise.resolve()
     }
-
-    /** Looks up for binaries to sign using the given path pattern. */
-    static lookup = lookupHelper(path => new BinaryToSign(path))
-
-    /** Looks up for binaries to sign using the given path patterns. */
-    static lookupMany = lookupManyHelper(BinaryToSign.lookup)
 }
 
 // ==============================
@@ -253,7 +248,8 @@ class BinaryToSign implements Signable {
 
 /** Helper used to concisely define patterns for an archive to sign.
  *
- * Consists of pattern of the archive path and set of patterns for files to sign inside the archive.
+ * Consists of pattern of the archive path
+ * and set of patterns for files to sign inside the archive.
  */
 type ArchivePattern = [glob.Pattern, glob.Pattern[]]
 
@@ -265,13 +261,13 @@ async function globAbs(pattern: glob.Pattern, options?: glob.Options): Promise<s
 
 /** Glob patterns relative to a given base directory. Base directory is allowed to be a pattern as
  * well.
- **/
+ */
 async function globAbsIn(
     base: glob.Pattern,
     pattern: glob.Pattern,
     options?: glob.Options
 ): Promise<string[]> {
-    return globAbs(path.join(base, pattern), options)
+    return globAbs(pathModule.join(base, pattern), options)
 }
 
 /** Generate a lookup function for a given Signable type. */
@@ -305,8 +301,7 @@ async function rmRf(path: string) {
  * Get a new temporary directory. Caller is responsible for cleaning up the directory.
  */
 async function getTmpDir(prefix?: string) {
-    const ret = await fs.mkdtemp(path.join(os.tmpdir(), prefix ?? 'enso-signing-'))
-    return ret
+    return await fs.mkdtemp(pathModule.join(os.tmpdir(), prefix ?? 'enso-signing-'))
 }
 
 // ====================
@@ -322,10 +317,10 @@ interface Input extends SigningContext {
 /** Entry point, meant to be used from an afterSign Electron Builder's hook. */
 export default async function (context: Input) {
     console.log('Environment: ', process.env)
-    const { appOutDir, productFilename, entitlements } = context
-    const appDir = path.join(appOutDir, `${productFilename}.app`)
-    const contentsDir = path.join(appDir, 'Contents')
-    const resourcesDir = path.join(contentsDir, 'Resources')
+    const { appOutDir, productFilename } = context
+    const appDir = pathModule.join(appOutDir, `${productFilename}.app`)
+    const contentsDir = pathModule.join(appDir, 'Contents')
+    const resourcesDir = pathModule.join(contentsDir, 'Resources')
 
     // Sign archives.
     console.log('Signing GraalVM elemenets...')
@@ -335,6 +330,8 @@ export default async function (context: Input) {
     for (const signable of await ensoPackageSignables(resourcesDir)) await signable.sign(context)
 
     // Finally re-sign the top-level enso.
-    const topLevelExecutable = new BinaryToSign(path.join(contentsDir, 'MacOS', productFilename))
+    const topLevelExecutable = new BinaryToSign(
+        pathModule.join(contentsDir, 'MacOS', productFilename)
+    )
     await topLevelExecutable.sign(context)
 }
