@@ -1,5 +1,5 @@
 /**
- * This script is for watching the whole IDE and spawning the electron process.
+ * @file This script is for watching the whole IDE and spawning the electron process.
  *
  * It sets up watchers for the client and content, and spawns the electron process with the IDE.
  * The spawned electron process can then use its refresh capability to pull the latest changes
@@ -9,18 +9,16 @@
  * To stop, use Ctrl+C.
  */
 
-import child_process from 'node:child_process'
-import path from 'node:path'
-import process from 'node:process'
-import fs from 'node:fs/promises'
+import * as childProcess from 'node:child_process'
+import * as fs from 'node:fs/promises'
+import * as path from 'node:path'
+import * as process from 'node:process'
 
-import esbuild from 'esbuild'
-import * as assert from 'assert'
+import * as esbuild from 'esbuild'
 
 import * as clientBundler from './esbuild-config.js'
 import * as contentBundler from '../content/esbuild-config.js'
-import * as esbuildWatch from '../../esbuild-watch.js'
-import { getIdeDirectory, getProjectManagerBundle, project_manager_bundle } from './paths.js'
+import * as paths from './paths.js'
 
 /** Set of esbuild watches for the client and content. */
 interface Watches {
@@ -28,47 +26,55 @@ interface Watches {
     content: esbuild.BuildResult
 }
 
-const ideDist = getIdeDirectory()
-const projectManagerBundle = getProjectManagerBundle()
+const IDE_DIR_PATH = paths.getIdeDirectory()
+const PROJECT_MANAGER_BUNDLE_PATH = paths.getProjectManagerBundlePath()
 
 console.log('Cleaning IDE dist directory.')
-await fs.rm(ideDist, { recursive: true, force: true })
-await fs.mkdir(ideDist, { recursive: true })
+await fs.rm(IDE_DIR_PATH, { recursive: true, force: true })
+await fs.mkdir(IDE_DIR_PATH, { recursive: true })
 
-const bothBundlesReady = new Promise<Watches>(async (resolve, reject) => {
-    console.log('Bundling client.')
-    const clientBundlerOpts: esbuild.BuildOptions = {
-        ...clientBundler.bundlerOptionsFromEnv(),
-        outdir: path.resolve(ideDist),
-        watch: {
-            onRebuild(error, result) {
-                if (error) {
-                    // We cannot carry on if the client failed to build, because electron executable
-                    // would immediately exit with an error.
-                    console.error('Client watch bundle failed:', error)
-                    reject(error)
-                } else {
-                    console.log('Client bundle updated.')
-                }
+const BOTH_BUNDLES_READY = new Promise<Watches>((resolve, reject) => {
+    void (async () => {
+        console.log('Bundling client.')
+        const clientBundlerOpts: esbuild.BuildOptions = {
+            ...clientBundler.bundlerOptionsFromEnv(),
+            outdir: path.resolve(IDE_DIR_PATH),
+            watch: {
+                onRebuild(error) {
+                    if (error) {
+                        // We cannot carry on if the client failed to build, because electron executable
+                        // would immediately exit with an error.
+                        console.error('Client watch bundle failed:', error)
+                        reject(error)
+                    } else {
+                        console.log('Client bundle updated.')
+                    }
+                },
             },
-        },
-    }
-    let client = await esbuild.build(clientBundlerOpts)
-    console.log('Result of client bundling: ', client)
+        }
+        const client = await esbuild.build(clientBundlerOpts)
+        console.log('Result of client bundling: ', client)
 
-    console.log('Bundling content.')
-    const contentOpts = contentBundler.watchOptions(() => console.log('Content bundle updated.'))
-    contentOpts.outdir = path.resolve(ideDist, 'assets')
-    const content = await esbuild.build(contentOpts)
-    console.log('Result of content bundling: ', content)
-    resolve({ client, content })
+        console.log('Bundling content.')
+        const contentOpts = contentBundler.watchOptions(() => {
+            console.log('Content bundle updated.')
+        })
+        contentOpts.outdir = path.resolve(IDE_DIR_PATH, 'assets')
+        const content = await esbuild.build(contentOpts)
+        console.log('Result of content bundling: ', content)
+        resolve({ client, content })
+    })()
 })
 
-const watches = await bothBundlesReady
+await BOTH_BUNDLES_READY
 console.log('Exposing Project Manager bundle.')
-await fs.symlink(projectManagerBundle, path.join(ideDist, project_manager_bundle), 'dir')
+await fs.symlink(
+    PROJECT_MANAGER_BUNDLE_PATH,
+    path.join(IDE_DIR_PATH, paths.PROJECT_MANAGER_BUNDLE),
+    'dir'
+)
 
-const electronArgs = [path.join(ideDist, 'index.cjs'), '--', ...process.argv.slice(2)]
+const ELECTRON_ARGS = [path.join(IDE_DIR_PATH, 'index.cjs'), '--', ...process.argv.slice(2)]
 
 process.on('SIGINT', () => {
     console.log('SIGINT received. Exiting.')
@@ -77,9 +83,9 @@ process.on('SIGINT', () => {
     process.exit(0)
 })
 
-while (true) {
+for (;;) {
     console.log('Spawning Electron process.')
-    const electronProcess = child_process.spawn('electron', electronArgs, {
+    const electronProcess = childProcess.spawn('electron', ELECTRON_ARGS, {
         stdio: 'inherit',
         shell: true,
     })
