@@ -1,38 +1,48 @@
 /** @file Configuration options of the application content (the web part). */
 
 import * as semver from 'semver'
-import { log, config, App } from '../../../../../target/ensogl-pack/linked-dist/index'
-import buildCfg from '../../../build.json'
 
-const Option = config.Option
-const Group = config.Group
-const logger = log.logger
-export { Option, Group, logger }
-export type Option<T> = config.Option<T>
+import * as linkedDist from '../../../../../target/ensogl-pack/linked-dist/index'
+import BUILD_INFO from '../../../build.json' assert { type: 'json' }
+
+// Aliases with the same name as the original.
+// eslint-disable-next-line no-restricted-syntax
+export const Option = linkedDist.config.Option
+// eslint-disable-next-line no-restricted-syntax
+export const Group = linkedDist.config.Group
+export const logger = linkedDist.log.logger
+// Declaring type with the same name as a variable.
+// eslint-disable-next-line @typescript-eslint/no-redeclare
+export type Option<T> = linkedDist.config.Option<T>
 
 // ===============
 // === Version ===
-// ===============s
+// ===============
 
-export class Version {
+export const VERSION = {
     /// Development version.
-    static dev = new semver.SemVer('0.0.0')
-    static devPrerelease = 'dev'
+    dev: new semver.SemVer('0.0.0'),
+    devPrerelease: 'dev',
 
     /// Version of the `client` js package.
-    static ide = new semver.SemVer(buildCfg.version, { loose: true })
+    ide: new semver.SemVer(BUILD_INFO.version, { loose: true }),
 
-    static isDev(): boolean {
-        const clientVersion = Version.ide
-        const releaseDev = clientVersion.compareMain(Version.dev) === 0
-        const prereleaseDev = clientVersion.prerelease.toString().includes(Version.devPrerelease)
+    isDev(): boolean {
+        const clientVersion = VERSION.ide
+        const releaseDev = clientVersion.compareMain(VERSION.dev) === 0
+        const prereleaseDev = clientVersion.prerelease.toString().includes(VERSION.devPrerelease)
         return releaseDev || prereleaseDev
-    }
+    },
 }
 
 // ===============
 // === Options ===
 // ===============
 
-import * as jsonCfg from './config.json'
-export const options = config.options.merge(config.objectToGroup(jsonCfg, { Version }))
+import CONFIG from './config.json' assert { type: 'json' }
+
+export const OPTIONS = linkedDist.config.options.merge(
+    // `valueEval` in `config.json` uses PascalCase scope variables for some reason.
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    linkedDist.config.objectToGroup(CONFIG, { Version: VERSION })
+)
