@@ -1,34 +1,32 @@
 /**
- * Configuration for the esbuild bundler and build/watch commands.
+ * @file Configuration for the esbuild bundler and build/watch commands.
  *
  * The bundler processes each entry point into a single file, each with no external dependencies and
  * minified. This primarily involves resolving all imports, along with some other transformations
  * (like TypeScript compilation).
  *
  * See the bundlers documentation for more information:
- * https://esbuild.github.io/getting-started/#bundling-for-node
+ * https://esbuild.github.io/getting-started/#bundling-for-node.
  */
 
-import fs from 'node:fs'
-import path, { dirname } from 'node:path'
-import child_process from 'node:child_process'
-import { fileURLToPath } from 'node:url'
+import * as childProcess from 'node:child_process'
+import * as fs from 'node:fs'
+import * as path from 'node:path'
+import * as url from 'node:url'
 
-import esbuild from 'esbuild'
-import plugin_yaml from 'esbuild-plugin-yaml'
-import { NodeModulesPolyfillPlugin } from '@esbuild-plugins/node-modules-polyfill'
-import { NodeGlobalsPolyfillPlugin } from '@esbuild-plugins/node-globals-polyfill'
-import aliasPlugin from 'esbuild-plugin-alias'
-// @ts-ignore
-import timePlugin from 'esbuild-plugin-time'
-// @ts-ignore
-import * as copy_plugin from 'enso-copy-plugin'
+import * as esbuild from 'esbuild'
+import * as esbuildPluginCopy from 'enso-copy-plugin'
+import * as esbuildPluginNodeGlobals from '@esbuild-plugins/node-globals-polyfill'
+import * as esbuildPluginNodeModules from '@esbuild-plugins/node-modules-polyfill'
+import esbuildPluginAlias from 'esbuild-plugin-alias'
+import esbuildPluginTime from 'esbuild-plugin-time'
+import esbuildPluginYaml from 'esbuild-plugin-yaml'
 
-import { require_env } from '../../utils.js'
-import * as BUILD_INFO from '../../build.json' assert { type: 'json' }
-import * as esbuildWatchHelper from '../../esbuild-watch.js'
+import * as esbuildWatch from '../../esbuild-watch.js'
+import * as utils from '../../utils.js'
+import BUILD_INFO from '../../build.json' assert { type: 'json' }
 
-export const thisPath = path.resolve(dirname(fileURLToPath(import.meta.url)))
+export const THIS_PATH = path.resolve(path.dirname(url.fileURLToPath(import.meta.url)))
 
 // =============================
 // === Environment variables ===
@@ -36,21 +34,24 @@ export const thisPath = path.resolve(dirname(fileURLToPath(import.meta.url)))
 
 export interface Arguments {
     /** List of files to be copied from WASM artifacts. */
-    wasm_artifacts: string
+    wasmArtifacts: string
     /** Directory with assets. Its contents are to be copied. */
-    assets_path: string
+    assetsPath: string
     /** Path where bundled files are output. */
-    output_path: string
+    outputPath: string
     /** The main JS bundle to load WASM and JS wasm-pack bundles. */
-    ensogl_app_path: string
+    ensoglAppPath: string
 }
 
+/**
+ * Get arguments from the environment.
+ */
 export function argumentsFromEnv(): Arguments {
-    const wasm_artifacts = require_env('ENSO_BUILD_GUI_WASM_ARTIFACTS')
-    const assets_path = require_env('ENSO_BUILD_GUI_ASSETS')
-    const output_path = path.resolve(require_env('ENSO_BUILD_GUI'), 'assets')
-    const ensogl_app_path = require_env('ENSO_BUILD_GUI_ENSOGL_APP')
-    return { wasm_artifacts, assets_path, output_path, ensogl_app_path }
+    const wasmArtifacts = utils.requireEnv('ENSO_BUILD_GUI_WASM_ARTIFACTS')
+    const assetsPath = utils.requireEnv('ENSO_BUILD_GUI_ASSETS')
+    const outputPath = path.resolve(utils.requireEnv('ENSO_BUILD_GUI'), 'assets')
+    const ensoglAppPath = utils.requireEnv('ENSO_BUILD_GUI_ENSOGL_APP')
+    return { wasmArtifacts, assetsPath, outputPath, ensoglAppPath }
 }
 
 // ===================
@@ -59,13 +60,13 @@ export function argumentsFromEnv(): Arguments {
 
 /**
  * Get output of a git command.
- * @param command Command line following the `git` program.
+ * @param command - Command line following the `git` program.
  * @returns Output of the command.
  */
 function git(command: string): string {
     // TODO [mwu] Eventually this should be removed, data should be provided by the build script through `BUILD_INFO`.
     //            The bundler configuration should not invoke git, it is not its responsibility.
-    return child_process.execSync(`git ${command}`, { encoding: 'utf8' }).trim()
+    return childProcess.execSync(`git ${command}`, { encoding: 'utf8' }).trim()
 }
 
 // ==============================
@@ -75,24 +76,25 @@ function git(command: string): string {
 /**
  * Static set of files that are always copied to the output directory.
  */
-export function always_copied_files(wasm_artifacts: string) {
+export function alwaysCopiedFiles(wasmArtifacts: string) {
     return [
-        path.resolve(thisPath, 'src', 'index.html'),
-        path.resolve(thisPath, 'src', 'run.js'),
-        path.resolve(thisPath, 'src', 'style.css'),
-        path.resolve(thisPath, 'src', 'docsStyle.css'),
-        ...wasm_artifacts.split(path.delimiter),
+        path.resolve(THIS_PATH, 'src', 'index.html'),
+        path.resolve(THIS_PATH, 'src', 'run.js'),
+        path.resolve(THIS_PATH, 'src', 'style.css'),
+        path.resolve(THIS_PATH, 'src', 'docsStyle.css'),
+        ...wasmArtifacts.split(path.delimiter),
     ]
 }
 
 /**
  * Generator that yields all files that should be copied to the output directory.
+ * @yields {string} The file path of the next file to be copied.
  */
-export async function* files_to_copy_provider(wasm_artifacts: string, assets_path: string) {
+export async function* filesToCopyProvider(wasmArtifacts: string, assetsPath: string) {
     console.log('Preparing a new generator for files to copy.')
-    yield* always_copied_files(wasm_artifacts)
-    for (const file of await fs.promises.readdir(assets_path)) {
-        yield path.resolve(assets_path, file)
+    yield* alwaysCopiedFiles(wasmArtifacts)
+    for (const file of await fs.promises.readdir(assetsPath)) {
+        yield path.resolve(assetsPath, file)
     }
     console.log('Generator for files to copy finished.')
 }
@@ -101,23 +103,29 @@ export async function* files_to_copy_provider(wasm_artifacts: string, assets_pat
 // === Bundling ===
 // ================
 
+/**
+ * Generate the builder options.
+ */
 export function bundlerOptions(args: Arguments): esbuild.BuildOptions {
-    const { output_path, ensogl_app_path, wasm_artifacts, assets_path } = args
+    const { outputPath, ensoglAppPath, wasmArtifacts, assetsPath } = args
     return {
-        absWorkingDir: thisPath,
+        /* eslint-disable @typescript-eslint/naming-convention */
+        absWorkingDir: THIS_PATH,
         bundle: true,
-        entryPoints: [path.resolve(thisPath, 'src', 'index.ts')],
-        outdir: output_path,
+        entryPoints: [path.resolve(THIS_PATH, 'src', 'index.ts')],
+        outdir: outputPath,
         outbase: 'src',
         plugins: [
-            plugin_yaml.yamlPlugin({}),
-            NodeModulesPolyfillPlugin(),
-            NodeGlobalsPolyfillPlugin({ buffer: true, process: true }),
-            aliasPlugin({ ensogl_app: ensogl_app_path }),
-            timePlugin(),
-            copy_plugin.create(() => files_to_copy_provider(wasm_artifacts, assets_path)),
+            esbuildPluginYaml.yamlPlugin({}),
+            esbuildPluginNodeModules.NodeModulesPolyfillPlugin(),
+            esbuildPluginNodeGlobals.NodeGlobalsPolyfillPlugin({ buffer: true, process: true }),
+            // We do not control naming of third-party options.
+            esbuildPluginAlias({ ensogl_app: ensoglAppPath }),
+            esbuildPluginTime(),
+            esbuildPluginCopy.create(() => filesToCopyProvider(wasmArtifacts, assetsPath)),
         ],
         define: {
+            // Disabling naming convention because these are third-party options.
             GIT_HASH: JSON.stringify(git('rev-parse HEAD')),
             GIT_STATUS: JSON.stringify(git('status --short --porcelain')),
             BUILD_INFO: JSON.stringify(BUILD_INFO),
@@ -144,20 +152,24 @@ export function bundlerOptions(args: Arguments): esbuild.BuildOptions {
             //    1 │ ...y{table.grow(1)}catch(err){if(!err instanceof RangeError){throw ...
             'suspicious-boolean-not': 'silent',
         },
+        /* eslint-enable @typescript-eslint/naming-convention */
     }
 }
 
+// This lint doesn't like "e.g." and "vs."".
+/* eslint-disable jsdoc/require-description-complete-sentence */
 /** The basic, common settings for the bundler, based on the environment variables.
  *
  * Note that they should be further customized as per the needs of the specific workflow (e.g. watch vs. build).
- **/
+ */
+/* eslint-enable jsdoc/require-description-complete-sentence */
 export function bundlerOptionsFromEnv(): esbuild.BuildOptions {
     return bundlerOptions(argumentsFromEnv())
 }
 
 /** ESBuild options for spawning a watcher, that will continuously rebuild the package. */
 export function watchOptions(onRebuild?: () => void, inject?: esbuild.BuildOptions['inject']) {
-    return esbuildWatchHelper.toWatchOptions(bundlerOptionsFromEnv(), onRebuild, inject)
+    return esbuildWatch.toWatchOptions(bundlerOptionsFromEnv(), onRebuild, inject)
 }
 
 /** ESBuild options for bundling (one-off build) the package.
