@@ -1,5 +1,4 @@
 /** @file This file is used to simply run the IDE. It can be not invoked if the IDE needs to be used
  * as a library. */
 
-/* eslint no-undef: "off" */
-window.enso.main()
+void window.enso.main()
