@@ -1,6 +1,7 @@
 /** @file This module exports wasm Rust glue code generated by EnsoGL Pack. */
 
-import { default as init } from 'wasm_rust_glue'
+import * as wasmRustGlue from 'wasm_rust_glue'
 
-/* eslint no-undef: "off" */
-exports.init = init
+// Eslint is not (and should not be) set up to check CommonJS.
+// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+exports.init = wasmRustGlue.default
