@@ -1,11 +1,15 @@
-import bundler from './esbuild-config.js'
-// @ts-ignore
-import * as server from 'enso-gui-server'
-import esbuild from 'esbuild'
+/** @file File watch and compile service. */
+import * as esbuild from 'esbuild'
 
-const opts = bundler.watchOptions(() => liveServer?.reload(), [server.LIVE_RELOAD_LISTENER_PATH])
-await esbuild.build(opts)
-const liveServer = await server.start({
-    root: opts.outdir,
-    assets: opts.outdir,
+import * as guiServer from 'enso-gui-server'
+
+import bundler from './esbuild-config.js'
+
+const OPTS = bundler.watchOptions(() => {
+    LIVE_SERVER.reload()
+}, [guiServer.LIVE_RELOAD_LISTENER_PATH])
+await esbuild.build(OPTS)
+const LIVE_SERVER = await guiServer.start({
+    root: OPTS.outdir,
+    assets: OPTS.outdir,
 })
